requirements.txt
sarvamai
sounddevice
scipy
numpy
boto3  
simpleaudio 
requests
soundfile
openai-whisper
chatterbox-tts
streamlit
elevenlabslib
bedrock-agentcore
bedrock-agentcore-starter-toolkit
assemblyai
pyaudio